Fixing failing tf tests
