+++ merlin_models/tf/prediction/item_prediction.py
@@ -20,15 +20,15 @@ import tensorflow as tf
 from tensorflow.python.layers.base import Layer
 from tensorflow.python.ops import embedding_ops
 
-from merlin_models.tf.block.transformations import L2Norm
-from merlin_models.tf.core import Block, EmbeddingWithMetadata, SequentialBlock
-from merlin_models.tf.prediction.sampling import InBatchSampler, ItemSampler, PopularityBasedSampler
-from merlin_models.utils.constants import MIN_FLOAT
 from merlin_standard_lib import Schema, Tag
 
+from ...utils.constants import MIN_FLOAT
 from ..block.aggregation import SequenceAggregation, SequenceAggregator
 from ..block.inputs import InputBlock
 from ..block.mlp import MLPBlock
+from ..block.transformations import L2Norm
+from ..core import Block, EmbeddingWithMetadata, SequentialBlock
+from ..prediction.sampling import InBatchSampler, ItemSampler, PopularityBasedSampler
 from ..typing import TabularData
 from .classification import CategFeaturePrediction, MultiClassClassificationTask
 from .ranking_metric import ranking_metrics
